pyob/set/classes/pyob_set.py
# ┌─────────────────────────────────────────────────────────────────────────────────────
# │ GENERAL IMPORTS
# └─────────────────────────────────────────────────────────────────────────────────────

from __future__ import annotations

from typing import Generic, TYPE_CHECKING, TypeVar

# ┌─────────────────────────────────────────────────────────────────────────────────────
# │ PROJECT IMPORTS
# └─────────────────────────────────────────────────────────────────────────────────────

from pyob.utils.classes.sequence import FrozenDict
from pyob.utils.mixins.pyobs import PyObs

if TYPE_CHECKING:
    from pyob.types import PyOb


# ┌─────────────────────────────────────────────────────────────────────────────────────
# │ PYOB SET
# └─────────────────────────────────────────────────────────────────────────────────────

# Define a PyOb TypeVar
PyObVar = TypeVar("PyObVar", bound="PyOb")


class PyObSet(PyObs[PyObVar], Generic[PyObVar]):
    """An abstract class for a collection of PyOb instances"""

    # ┌─────────────────────────────────────────────────────────────────────────────────
    # │ TYPE DECLARATION: COUNTS
    # └─────────────────────────────────────────────────────────────────────────────────

    # Declare type of counts
    _counts: FrozenDict[PyObVar, int]

    # ┌─────────────────────────────────────────────────────────────────────────────────
    # │ __ADD__
    # └─────────────────────────────────────────────────────────────────────────────────

    def __add__(self, other: PyObVar) -> None:
        """Add Method"""

        # Raise NotImplementedError
        raise NotImplementedError

    # ┌─────────────────────────────────────────────────────────────────────────────────
    # │ __INIT__
    # └─────────────────────────────────────────────────────────────────────────────────

    def __init__(self, _counts: dict[PyObVar, int] | None = None) -> None:
        """Init Method"""

        # Call parent init method
        super().__init__(_counts=_counts)

        # Freeze the PyObSet counts dictionary
        # A PyObSet should be considered immutable
        self._counts = FrozenDict(self._counts)

    # ┌─────────────────────────────────────────────────────────────────────────────────
    # │ __SUB__
    # └─────────────────────────────────────────────────────────────────────────────────

    def __sub__(self, other: PyObVar) -> None:
        """Subtract Method"""

        # Raise NotImplementedError
        raise NotImplementedError
